Clarify logic around dev builds and rebuilding
The concepts of using the dev builds for nitro and blockscout and
rebuilding are now split into different variables internally to the
test-node.bash script and can be controlled independently.

The --dev flag by itself means to use the dev builds and to rebuild
them. --no-build-dev-nitro and -no-build-dev-blockscout disable
rebuilding the respective dev images. --no-build disables rebuilding the
dev images and also rebuilding contracts related images, and the node
images. This finer grained control saves time in the development cycle
by allowing the user to rebuild only what is needed.

An example command line to reinitialize the blockchain, use the dev
images, but not rebuild anything except blockscout would be:
./test-node.bash --init --dev --no-build --build-dev-blockscout --blockscout

Flags are read from left to right; flags to the right cancel
the effects of flags to the left.